% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFormulaMI.R
\name{gFormulaAnalyse}
\alias{gFormulaAnalyse}
\title{Analyse a set of gFormulaMI synthetic imputed datasets}
\usage{
gFormulaAnalyse(fits)
}
\arguments{
\item{fits}{Collection of model fits produced by a call of the form with(imps, lm(y~regime))}
}
\description{
Analyse a set of gFormulaMI synthetic imputed datasets
}
