% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFormulaMI.R
\name{gFormulaImpute}
\alias{gFormulaImpute}
\title{G-formula multiple imputation}
\usage{
gFormulaImpute(
  data,
  M = 50,
  trtVarStem,
  timePoints,
  trtRegimes,
  nSim = NULL,
  micePrintFlag = FALSE,
  method = NULL,
  predictorMatrix = NULL
)
}
\arguments{
\item{data}{The observed data frame}

\item{M}{The number of imputed datasets to generate}

\item{trtVarStem}{String indicating the stem of the treatment variables}

\item{timePoints}{Integer specifying the number of timepoints (including baseline)}

\item{trtRegimes}{A vector specifying the treatment regime of interest, or a list of
vectors specifying the treatment regimes of interest}

\item{nSim}{The number of individuals to simulate in each imputed dataset. Defaults to
number of individuals in observed data}

\item{micePrintFlag}{TRUE/FALSE specifying whether the output from the call to mice
should be printed}

\item{method}{An optional method argument to pass to mice. If not specified, the default
is to impute continuous variables using normal linear regression (norm), binary variables using
logistic regression (logreg), polytomous regression for unordered factors and
proportional odds model for ordered factors}

\item{predictorMatrix}{An optional predictor matrix to specify which variables to use
as predictors in the imputation models. The default is to impute sequentially, i.e. impute
using all variables to the left of the variable being imputed as covariates}
}
\value{
an S3 object of class mids (multiply imputed dataset)
}
\description{
\code{gFormulaImpute} creates multiple imputed synthetic datasets of longitudinal histories under
specified treatment regimes of interest, based on the G-formula.
}
\details{
\code{gFormulaImpute} creates multiple imputed synthetic datasets of longitudinal histories under
specified treatment regimes of interest, based on the G-formula. Specifically, to
the observed data frame, an additional \code{nSim} rows are added in which all variables are set to
missing, except the time-varying treatment variables. The latter are set to the values
as specified in the \code{trtRegimes} argument. If multiple treatment regimes are specified,
\code{nSim} rows are added for each of the treatment regimes.

\code{gFormulaImpute} uses the \code{mice} package to impute the potential outcome values of the
time-varying confounders and outcome in the synthetic datasets. Imputation is performed
sequentially from left to right in the data frame. As such, the variables must be ordered
in time in the input data frame, with the time-varying confounders at each time followed
by the corresponding treatment variable at that time.

For the data argument, \code{gFormulaImpute} expects either a fully observed (complete) data frame,
or else a set of multiple imputation stored in an object of class mids, created by mice
in the mice package.

\code{gFormulaImpute} returns an object of class \code{mids}. This can be analysed using the same
methods that imputed datasets from \code{mice} can be analysed with. However, Rubin's standard
pooling rules are not valid for analysis of the synthetic datasets. Instead, the synthetic
variance estimator of Raghunathan et al (2003) must be used, as implemented in the
\code{gFormulaAnalyse} function.
}
